requirements.txt
openpyxl~=3.1.4
lxml
bs4~=0.0.2
requests~=2.32.3
aiogram~=3.8.0
python-dotenv~=1.0.1
asyncio~=3.4.3
SQLAlchemy~=2.0.31
beautifulsoup4~=4.12.3
aiosqlite